Move base content model, house and pskyer
