+++ gyrinx/content/models/content.py
@@ -24,8 +24,10 @@ from polymorphic.models import PolymorphicModel
 from simple_history.models import HistoricalRecords
 from simpleeval import simple_eval
 
+from gyrinx.content.models.psyker import ContentPsykerDiscipline, ContentPsykerPower
 from gyrinx.core.models.base import AppBase
 from gyrinx.content.models.base import Content
+from gyrinx.content.models.house import ContentHouse
 from gyrinx.content.models.skill import ContentSkill, ContentSkillCategory
 from gyrinx.core.models.util import ModContext
 from gyrinx.models import (
@@ -39,101 +41,6 @@ from gyrinx.models import (
 from gyrinx.tracker import track
 
 logger = logging.getLogger(__name__)
-
-##
-## Content Models
-##
-
-
-class ContentHouse(Content):
-    """
-    Represents a faction or house that fighters can belong to.
-    """
-
-    help_text = "The Content House identifies the house or faction of a fighter."
-    name = models.CharField(max_length=255, db_index=True)
-    skill_categories = models.ManyToManyField(
-        "ContentSkillCategory",
-        blank=True,
-        related_name="houses",
-        verbose_name="Unique Skill Categories",
-    )
-    generic = models.BooleanField(
-        default=False,
-        help_text="If checked, fighters in this House can join lists and gangs of any other House.",
-    )
-    legacy = models.BooleanField(
-        default=False,
-        help_text="If checked, this House is considered a legacy/older faction.",
-    )
-    can_hire_any = models.BooleanField(
-        default=False,
-        help_text="If checked, this House can hire any fighter from any house (except stash fighters).",
-    )
-    can_buy_any = models.BooleanField(
-        default=False,
-        help_text="If checked, this House can buy any equipment from any equipment list and trading post.",
-    )
-
-    history = HistoricalRecords()
-
-    def fighters(self):
-        """
-        Returns all fighters associated with this house.
-        """
-        return self.contentfighter_set.all()
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name = "House"
-        verbose_name_plural = "Houses"
-        ordering = ["name"]
-
-
-class ContentPsykerDiscipline(Content):
-    """
-    Represents a discipline of Psyker/Wyrd powers.
-    """
-
-    name = models.CharField(max_length=255, unique=True)
-    generic = models.BooleanField(
-        default=False,
-        help_text="If checked, this discipline can be used by any psyker.",
-    )
-    history = HistoricalRecords()
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name = "Psyker Discipline"
-        verbose_name_plural = "Psyker Disciplines"
-        ordering = ["name"]
-
-
-class ContentPsykerPower(Content):
-    """
-    Represents a specific power within a discipline of Psyker/Wyrd powers.
-    """
-
-    name = models.CharField(max_length=255)
-    discipline = models.ForeignKey(
-        ContentPsykerDiscipline,
-        on_delete=models.CASCADE,
-        related_name="powers",
-    )
-    history = HistoricalRecords()
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name = "Psyker Power"
-        verbose_name_plural = "Psyker Powers"
-        ordering = ["discipline__name", "name"]
-        unique_together = ["name", "discipline"]
 
 
 class ContentFighterPsykerDisciplineAssignment(Content):
