+++ gyrinx/core/tests/test_psyker.py
@@ -4,10 +4,9 @@ from django.forms import ValidationError
 from gyrinx.content.models import (
     ContentFighterPsykerDisciplineAssignment,
     ContentFighterPsykerPowerDefaultAssignment,
-    ContentPsykerDiscipline,
-    ContentPsykerPower,
     ContentRule,
 )
+from gyrinx.content.models.psyker import ContentPsykerDiscipline, ContentPsykerPower
 from gyrinx.core.models.list import ListFighter, ListFighterPsykerPowerAssignment
